Changing variable name
The sensitivity variables include "missing_exposure_days" which was
calculated in a different manner than the "missing exposure day" count
used to rate the data quality of an exposure day.  I renamed the
sensitivity variable to "no_exposure_data_days", which means it is the
number of days in a time frame that have no daily exposure record data
at all.  This is different than assessing missing exposure days for
*each pollutant*.  In exposure assessments of daily exposure values,
missing exposures is with respect to a given type of pollutant whereas
no_exposure_data_days means no exposure values of any kind exist for
certain days in a time frame.