Experimenting with GitHub's "docs" folder to support web site
GitHub lets you either create a separate repository space for your web
site, or you can put a "docs" folder in your main repository and
anything that gets put in there is then used to support a project web
site.  Ideally, I'd like to find an option for specifying any folder
within the repository -- such as the docs folder within the Eclipse
project.  But for now, we'll copy the files to the directory the Github
documentation I'm ready says to use.